# server/captured-shop-server/ent/user.go
// Code generated by ent, DO NOT EDIT.

package ent

import (
	"product-server/ent/user"
	"fmt"
	"strings"
	"time"

	"entgo.io/ent"
	"entgo.io/ent/dialect/sql"
)

// User is the model entity for the User schema.
type User struct {
	config `json:"-"`
	// ID of the ent.
	ID string `json:"id,omitempty"`
	// Email holds the value of the "email" field.
	Email string `json:"email,omitempty"`
	// Password holds the value of the "password" field.
	Password string `json:"password,omitempty"`
	// KrName holds the value of the "kr_name" field.
	KrName string `json:"kr_name,omitempty"`
	// RegisterAt holds the value of the "register_at" field.
	RegisterAt time.Time `json:"register_at,omitempty"`
	// LastLogin holds the value of the "last_login" field.
	LastLogin time.Time `json:"last_login,omitempty"`
	// SignUpType holds the value of the "sign_up_type" field.
	SignUpType string `json:"sign_up_type,omitempty"`
	// Edges holds the relations/edges for other nodes in the graph.
	// The values are being populated by the UserQuery when eager-loading is set.
	Edges        UserEdges `json:"edges"`
	selectValues sql.SelectValues
}

// UserEdges holds the relations/edges for other nodes in the graph.
type UserEdges struct {
	// OrderHistories holds the value of the order_histories edge.
	OrderHistories []*OrderHistory `json:"order_histories,omitempty"`
	// UserAddress holds the value of the user_address edge.
	UserAddress []*UserAddress `json:"user_address,omitempty"`
	// loadedTypes holds the information for reporting if a
	// type was loaded (or requested) in eager-loading or not.
	loadedTypes [2]bool
}

// OrderHistoriesOrErr returns the OrderHistories value or an error if the edge
// was not loaded in eager-loading.
func (e UserEdges) OrderHistoriesOrErr() ([]*OrderHistory, error) {
	if e.loadedTypes[0] {
		return e.OrderHistories, nil
	}
	return nil, &NotLoadedError{edge: "order_histories"}
}

// UserAddressOrErr returns the UserAddress value or an error if the edge
// was not loaded in eager-loading.
func (e UserEdges) UserAddressOrErr() ([]*UserAddress, error) {
	if e.loadedTypes[1] {
		return e.UserAddress, nil
	}
	return nil, &NotLoadedError{edge: "user_address"}
}

// scanValues returns the types for scanning values from sql.Rows.
func (*User) scanValues(columns []string) ([]any, error) {
	values := make([]any, len(columns))
	for i := range columns {
		switch columns[i] {
		case user.FieldID, user.FieldEmail, user.FieldPassword, user.FieldKrName, user.FieldSignUpType:
			values[i] = new(sql.NullString)
		case user.FieldRegisterAt, user.FieldLastLogin:
			values[i] = new(sql.NullTime)
		default:
			values[i] = new(sql.UnknownType)
		}
	}
	return values, nil
}

// assignValues assigns the values that were returned from sql.Rows (after scanning)
// to the User fields.
func (u *User) assignValues(columns []string, values []any) error {
	if m, n := len(values), len(columns); m < n {
		return fmt.Errorf("mismatch number of scan values: %d != %d", m, n)
	}
	for i := range columns {
		switch columns[i] {
		case user.FieldID:
			if value, ok := values[i].(*sql.NullString); !ok {
				return fmt.Errorf("unexpected type %T for field id", values[i])
			} else if value.Valid {
				u.ID = value.String
			}
		case user.FieldEmail:
			if value, ok := values[i].(*sql.NullString); !ok {
				return fmt.Errorf("unexpected type %T for field email", values[i])
			} else if value.Valid {
				u.Email = value.String
			}
		case user.FieldPassword:
			if value, ok := values[i].(*sql.NullString); !ok {
				return fmt.Errorf("unexpected type %T for field password", values[i])
			} else if value.Valid {
				u.Password = value.String
			}
		case user.FieldKrName:
			if value, ok := values[i].(*sql.NullString); !ok {
				return fmt.Errorf("unexpected type %T for field kr_name", values[i])
			} else if value.Valid {
				u.KrName = value.String
			}
		case user.FieldRegisterAt:
			if value, ok := values[i].(*sql.NullTime); !ok {
				return fmt.Errorf("unexpected type %T for field register_at", values[i])
			} else if value.Valid {
				u.RegisterAt = value.Time
			}
		case user.FieldLastLogin:
			if value, ok := values[i].(*sql.NullTime); !ok {
				return fmt.Errorf("unexpected type %T for field last_login", values[i])
			} else if value.Valid {
				u.LastLogin = value.Time
			}
		case user.FieldSignUpType:
			if value, ok := values[i].(*sql.NullString); !ok {
				return fmt.Errorf("unexpected type %T for field sign_up_type", values[i])
			} else if value.Valid {
				u.SignUpType = value.String
			}
		default:
			u.selectValues.Set(columns[i], values[i])
		}
	}
	return nil
}

// Value returns the ent.Value that was dynamically selected and assigned to the User.
// This includes values selected through modifiers, order, etc.
func (u *User) Value(name string) (ent.Value, error) {
	return u.selectValues.Get(name)
}

// QueryOrderHistories queries the "order_histories" edge of the User entity.
func (u *User) QueryOrderHistories() *OrderHistoryQuery {
	return NewUserClient(u.config).QueryOrderHistories(u)
}

// QueryUserAddress queries the "user_address" edge of the User entity.
func (u *User) QueryUserAddress() *UserAddressQuery {
	return NewUserClient(u.config).QueryUserAddress(u)
}

// Update returns a builder for updating this User.
// Note that you need to call User.Unwrap() before calling this method if this User
// was returned from a transaction, and the transaction was committed or rolled back.
func (u *User) Update() *UserUpdateOne {
	return NewUserClient(u.config).UpdateOne(u)
}

// Unwrap unwraps the User entity that was returned from a transaction after it was closed,
// so that all future queries will be executed through the driver which created the transaction.
func (u *User) Unwrap() *User {
	_tx, ok := u.config.driver.(*txDriver)
	if !ok {
		panic("ent: User is not a transactional entity")
	}
	u.config.driver = _tx.drv
	return u
}

// String implements the fmt.Stringer.
func (u *User) String() string {
	var builder strings.Builder
	builder.WriteString("User(")
	builder.WriteString(fmt.Sprintf("id=%v, ", u.ID))
	builder.WriteString("email=")
	builder.WriteString(u.Email)
	builder.WriteString(", ")
	builder.WriteString("password=")
	builder.WriteString(u.Password)
	builder.WriteString(", ")
	builder.WriteString("kr_name=")
	builder.WriteString(u.KrName)
	builder.WriteString(", ")
	builder.WriteString("register_at=")
	builder.WriteString(u.RegisterAt.Format(time.ANSIC))
	builder.WriteString(", ")
	builder.WriteString("last_login=")
	builder.WriteString(u.LastLogin.Format(time.ANSIC))
	builder.WriteString(", ")
	builder.WriteString("sign_up_type=")
	builder.WriteString(u.SignUpType)
	builder.WriteByte(')')
	return builder.String()
}

// Users is a parsable slice of User.
type Users []*User